refactor(contact-service): userId类型 int→Long (Service层)

+++ AI-Shopping-backend_Eureka/contact-service/src/main/java/com/gzasc/aishopping/contact/service/UserContactService.java
@@ -5,11 +5,11 @@ import com.gzasc.aishopping.contact.model.Contact;
 import java.util.List;
 
 public interface UserContactService {
-    int createContact(Contact contact, int userId);
-    int deleteContact(int id, int userId);
-    int updateContact(Contact contact, int userId);
-    List<Contact> getContactsByUserId(int userId);
-    int setDefaultContact(int id, int userId);
+    int createContact(Contact contact, Long userId);
+    int deleteContact(int id, Long userId);
+    int updateContact(Contact contact, Long userId);
+    List<Contact> getContactsByUserId(Long userId);
+    int setDefaultContact(int id, Long userId);
     // 内部微服务专属
     Contact g(int id);
 }
